+++ halomod/__init__.py
@@ -1,3 +1,3 @@
-__version__ = "1.4.2"
+__version__ = "1.4.3"
 from halo_model import HaloModel
 from integrate_corr import ProjectedCF, AngularCF