+++ halomod/integrate_corr.py
@@ -13,6 +13,7 @@ from halo_model import HaloModel
 from hmf._cache import cached_property, parameter
 from halo_exclusion import dblsimps
 from hmf.cosmo import Cosmology as csm
+import warnings
 
 class ProjectedCF(HaloModel):
     def __init__(self, rp_min=0.01, rp_max=50.0, rp_num=30, rp_log=True, proj_limit=None, **kwargs):
@@ -210,6 +211,9 @@ class AngularCF(HaloModel):
                  logu_min=-4,logu_max=2.3,unum=100,check_p_norm=True, p_of_z=True,
                  **kwargs):
         super(AngularCF, self).__init__(**kwargs)
+
+        if self.z < zmin or self.z>zmax:
+            warnings.warn("Your specified redshift (z=%s) is not within your selection function, z=(%s,%s)"%(self.z,zmin,zmax))
 
         if p1 is None:
             p1 = flat_z_dist(zmin,zmax)