+++ CHANGELOG.rst
@@ -3,18 +3,40 @@ Releases
 
 Development Version
 ~~~~~~~~~~~~~~~~~~~
-Features
-++++++++
-
-Enhancements
-++++++++++++
-
-Bugfixes
-++++++++
 
 
 Older Versions
 ~~~~~~~~~~~~~~
+v1.4.3
+------
+23rd September, 2016
+
+Features
+++++++++
+* Function to populate a halo catalogue with HOD-derived galaxies now properly implemented.
+* Populate routine also returns an array with indices of the halos associated with each galaxy.
+
+Enhancements
+++++++++++++
+* When matching mean galaxy density to given value, the first guess is now the DM Mmin, rather than arbitrary 8.
+* Better error message for NGException
+* mean_gal_den now returns the *calculated* mean density, rather than ng if it exists.
+* If supplied redshift is outside redshift selection for AngularCF, warning is printed.
+* Mmin now set to 0 by default, to enable better matter-matter statistics.
+* Entirely revised system for HODs, especially concerning the "central condition":
+  * Ns, not Ntot, now modified, giving consistent results for all derived quantities
+  * Pair-counts now intrinsic to HOD class.
+  * ``populate()`` routine handles both cases -- where centrals are required, and not.
+  * Documentation in HOD module explaining the assumptions made.
+
+Bugfixes
+++++++++
+* __density_mod_mm added so that __density_mod not overwritten when getting matter correlations.
+* __density_mod_mm modified to account for the fact that m[0] != 0, when halo exclusion performed.
+* Several fixes for correlation functions not being counts (+1 errors)
+
+
+
 v1.4.2
 ------
 2nd September, 2016